+++ blocco esercizi 2/UnitTestBlocco2/es05Test.cs
@@ -10,19 +10,19 @@ namespace UnitTestBlocco2
         [TestMethod]
         public void TestEs05Factor()
         {
-            float factor = Program.factorial(5);
+            double factor = Program.factorial(5);
             Assert.AreEqual(factor, 120);
         }
         [TestMethod]
         public void TestEs05Exp()
         {
-            float exp = Program.exponential(2, 5);
+            double exp = Program.exponential(2, 5);
             Assert.AreEqual(exp, 32);
         }
         [TestMethod]
         public void TestEs05Sum()
         {
-            float sum = Program.calculate(1, 3);
+            double sum = Program.calculate(1, 3);
             Assert.AreEqual(sum, 10);
 
         }
